requirements.txt
transformers
torch
python-dotenv
langchain
langchain-huggingface
langgraph
streamlit
accelerate
fastapi
uvicorn
pydantic
faiss-cpu
numpy
